split KyuubiExtensionSuite to different extension ut class
